+++ doc/source/library/pymodbus.client.async.factory.rst
@@ -1,7 +1,7 @@
-pymodbus\.client\.async\.factory package
+pymodbus\.client\.asynchronous\.factory package
 ========================================
 
-.. automodule:: pymodbus.client.async.factory
+.. automodule:: pymodbus.client.asynchronous.factory
     :members:
     :undoc-members:
     :show-inheritance:
@@ -9,26 +9,26 @@ pymodbus\.client\.async\.factory package
 Submodules
 ----------
 
-pymodbus\.client\.async\.factory\.serial module
+pymodbus\.client\.asynchronous\.factory\.serial module
 -----------------------------------------------
 
-.. automodule:: pymodbus.client.async.factory.serial
+.. automodule:: pymodbus.client.asynchronous.factory.serial
     :members:
     :undoc-members:
     :show-inheritance:
 
-pymodbus\.client\.async\.factory\.tcp module
+pymodbus\.client\.asynchronous\.factory\.tcp module
 --------------------------------------------
 
-.. automodule:: pymodbus.client.async.factory.tcp
+.. automodule:: pymodbus.client.asynchronous.factory.tcp
     :members:
     :undoc-members:
     :show-inheritance:
 
-pymodbus\.client\.async\.factory\.udp module
+pymodbus\.client\.asynchronous\.factory\.udp module
 --------------------------------------------
 
-.. automodule:: pymodbus.client.async.factory.udp
+.. automodule:: pymodbus.client.asynchronous.factory.udp
     :members:
     :undoc-members:
     :show-inheritance:
