+++ doc/source/library/pymodbus.client.async.schedulers.rst
@@ -1,7 +1,7 @@
-pymodbus\.client\.async\.schedulers package
+pymodbus\.client\.asynchronous\.schedulers package
 ===========================================
 
-.. automodule:: pymodbus.client.async.schedulers
+.. automodule:: pymodbus.client.asynchronous.schedulers
     :members:
     :undoc-members:
     :show-inheritance:
