+++ examples/functional/asynchronous_ascii_client.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 import unittest
-from pymodbus.client.async import ModbusSerialClient as ModbusClient
+from pymodbus.client.asynchronous  import ModbusSerialClient as ModbusClient
 from base_runner import Runner
 
 class AsynchronousAsciiClient(Runner, unittest.TestCase):
