+++ examples/gui/gui_common.py
@@ -12,7 +12,7 @@ from threading import Thread
 # -------------------------------------------------------------------------- #
 from twisted.internet import reactor
 from twisted.internet import error as twisted_error
-from pymodbus.server.async import ModbusServerFactory
+from pymodbus.server.asynchronous import ModbusServerFactory
 from pymodbus.datastore import ModbusServerContext,ModbusSlaveContext
 
 # -------------------------------------------------------------------------- #
