+++ examples/gui/tk/simulator.py
@@ -24,7 +24,7 @@ tksupport.install(root)
 # --------------------------------------------------------------------------- #
 from twisted.internet import reactor
 from twisted.internet import error as twisted_error
-from pymodbus.server.async import ModbusServerFactory
+from pymodbus.server.asynchronous import ModbusServerFactory
 from pymodbus.datastore import ModbusServerContext,ModbusSlaveContext
 
 #--------------------------------------------------------------------------#
