+++ pymodbus/client/asynchronous/factory/serial.py
@@ -6,8 +6,8 @@ from __future__ import absolute_import
 
 import logging
 
-from pymodbus.client.async import schedulers
-from pymodbus.client.async.thread import EventLoopThread
+from pymodbus.client.asynchronous  import schedulers
+from pymodbus.client.asynchronous.thread import EventLoopThread
 
 LOGGER = logging.getLogger(__name__)
 
@@ -66,7 +66,7 @@ def io_loop_factory(port=None, framer=None, **kwargs):
     """
 
     from tornado.ioloop import IOLoop
-    from pymodbus.client.async.tornado import (AsyncModbusSerialClient as
+    from pymodbus.client.asynchronous.tornado import (AsyncModbusSerialClient as
                                                Client)
 
     ioloop = IOLoop()
@@ -88,7 +88,7 @@ def async_io_factory(port=None, framer=None, **kwargs):
     :return: asyncio event loop and serial client
     """
     import asyncio
-    from pymodbus.client.async.asyncio import (ModbusClientProtocol,
+    from pymodbus.client.asynchronous.asyncio import (ModbusClientProtocol,
                                                AsyncioModbusSerialClient)
     loop = kwargs.pop("loop", None) or asyncio.get_event_loop()
     proto_cls = kwargs.pop("proto_cls", None) or ModbusClientProtocol
