+++ pymodbus/client/asynchronous/tornado/__init__.py
@@ -15,7 +15,7 @@ from tornado.ioloop import IOLoop
 from tornado.iostream import IOStream
 from tornado.iostream import BaseIOStream
 
-from pymodbus.client.async.mixins import (AsyncModbusClientMixin,
+from pymodbus.client.asynchronous.mixins import (AsyncModbusClientMixin,
                                           AsyncModbusSerialClientMixin)
 from pymodbus.exceptions import ConnectionException
 from pymodbus.compat import byte2int
